requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1
djangorestframework==3.13.1
dnspython==2.2.1
kombu==5.2.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
prompt-toolkit==3.0.30
pymongo==4.2.0
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
